[MOL-20717][RP] move VariantInternalProps to internal-types.ts
