+++ src/language-switcher/link-container-variant.tsx
@@ -7,7 +7,7 @@ import {
     LinkListItem,
 } from "./link-container-variant.style";
 import { ARIA_LABEL, LANGUAGE_CODES, LANGUAGE_DISPLAY_MAP } from "./data";
-import { VariantInternalProps } from "./types";
+import { VariantInternalProps } from "./internal-types";
 
 export const LinkContainerVariant = ({
     selectedLanguage,
